Moved all handling of compressed files to compressed data manager
